nvmap: Fix type casting issue
CERT-C is reporting following issue:
- Casting entry->sci_ipc_id from unsigned long long to unsigned int
without checking its value may result in lost or misinterpreted data.
Fix this by changing the data type of sid to u64.

CID 587909
Bug 3745813

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2816557
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/video/tegra/nvmap/nvmap_sci_ipc.c
@@ -38,7 +38,7 @@ struct nvmap_sci_ipc {
 
 struct free_sid_node {
 	struct list_head list;
-	u32 sid;
+	u64 sid;
 };
 
 /* An rb-tree root node for holding sci_ipc_id of clients */
